# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Sentinel AML API",
        "description": "Transaction monitoring: ingestion, rule-based detection, alerts and case management. Roles: ANALYST (queue & disposition), SUPERVISOR (+ full PII, case closure, reopen), ADMIN (+ ingestion, rule/watchlist/FX configuration).",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "basic": []
        }
    ],
...
    "paths": {
...
        "/api/v1/ingest/transactions": {
            "post": {
                "tags": [
                    "Ingestion"
                ],
                "summary": "Bulk load transactions from a JSON array, then run detection in parallel (detect=false to skip)",
                "operationId": "transactions",
                "parameters": [
                    {
                        "name": "detect",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/TransactionRecord"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestionResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/cases/{id}/transition": {
            "post": {
                "tags": [
                    "Cases"
                ],
                "summary": "Move a case: INVESTIGATING | SAR_FILED (SUPERVISOR) | CLOSED (SUPERVISOR; closes escalated alerts)",
                "description": "409 for a transition not allowed from the current status, 403 if the role is insufficient",
                "operationId": "transition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Transition"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/alerts/{id}/actions": {
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "START_REVIEW | CLOSE (disposition + reason) | ESCALATE (reason, optional caseId) | REOPEN (SUPERVISOR)",
                "description": "409 if the action is not allowed from the alert's current status; 422 if disposition/reason is missing",
                "operationId": "act",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/rules/{code}/history": {
            "get": {
                "tags": [
                    "Rules"
                ],
                "summary": "Every past version of a rule's configuration",
                "operationId": "history",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RuleConfigHistory"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Alert queue (PII masked), highest risk first by default",
                "operationId": "list_4",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "OPEN",
                                    "UNDER_REVIEW",
                                    "ESCALATED",
                                    "CLOSED"
                                ]
                            }
                        }
                    },
                    {
                        "name": "minScore",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "maximum": 100,
                            "minimum": 0,
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "ruleCode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customerId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "assignee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "maximum": 200,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32",
                            "default": 25
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "riskScore"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseAlertSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Alert detail: explanation per rule, evidence transactions, customer (PII visible to SUPERVISOR+)",
                "operationId": "get_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/alerts/{id}/history": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Immutable audit trail of this alert (creation, aggregation, every state transition)",
                "operationId": "history_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuditLog"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TransactionRecord": {
                "required": [
                    "accountNumber",
                    "amount",
                    "channel",
                    "currency",
                    "direction",
                    "occurredAt",
                    "txnRef"
                ],
                "type": "object",
                "properties": {
                    "txnRef": {
                        "maxLength": 60,
                        "minLength": 0,
                        "type": "string"
                    },
                    "accountNumber": {
                        "maxLength": 40,
                        "minLength": 0,
                        "type": "string"
                    },
                    "direction": {
                        "pattern": "CREDIT|DEBIT",
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "pattern": "[A-Z]{3}",
                        "type": "string"
                    },
                    "counterpartyName": {
                        "maxLength": 200,
                        "minLength": 0,
                        "type": "string"
                    },
                    "counterpartyAccount": {
                        "maxLength": 60,
                        "minLength": 0,
                        "type": "string"
                    },
                    "counterpartyCountry": {
                        "pattern": "[A-Z]{2}",
                        "type": "string"
                    },
                    "channel": {
                        "pattern": "CASH|WIRE|NEFT|RTGS|IMPS|UPI|CARD|ATM|CHEQUE|SWIFT",
                        "type": "string"
                    },
                    "jurisdiction": {
                        "pattern": "[A-Z]{2}",
                        "type": "string"
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "IngestionResult": {
                "type": "object",
                "properties": {
                    "batchId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "recordType": {
                        "type": "string"
                    },
                    "received": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "accepted": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "duplicates": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "rejected": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RowError"
                        }
                    },
                    "errorsTruncated": {
                        "type": "boolean"
                    },
                    "ingestMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "detection": {
                        "$ref": "#/components/schemas/Summary"
                    }
                }
            },
            "RowError": {
                "type": "object",
                "properties": {
                    "line": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "key": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                }
            },
            "Summary": {
                "type": "object",
                "properties": {
                    "evaluated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "skipped": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "transactionsWithHits": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "alertsCreated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "alertsUpdated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "failures": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "elapsedMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "txnPerSecond": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "AmlCase": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "caseRef": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "INVESTIGATING",
                            "SAR_FILED",
                            "CLOSED"
                        ]
                    },
                    "priority": {
                        "type": "string"
                    },
                    "assignee": {
                        "type": "string"
                    },
                    "outcome": {
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "rowVersion": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "AuditLog": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "entityType": {
                        "type": "string"
                    },
                    "entityId": {
                        "type": "string"
                    },
                    "action": {
                        "type": "string"
                    },
                    "fromState": {
                        "type": "string"
                    },
                    "toState": {
                        "type": "string"
                    },
                    "actor": {
                        "type": "string"
                    },
                    "details": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CaseDetail": {
                "type": "object",
                "properties": {
                    "caseInfo": {
                        "$ref": "#/components/schemas/AmlCase"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerView"
                    },
                    "alerts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertSummary"
                        }
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CaseNote"
                        }
                    },
                    "history": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuditLog"
                        }
                    }
                }
            },
            "CaseNote": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "caseId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "author": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CustomerView": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "externalRef": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "nationalId": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerType": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "kycRiskRating": {
                        "type": "string"
                    },
                    "piiMasked": {
                        "type": "boolean"
                    }
                }
            },
            "Transition": {
                "required": [
                    "status"
                ],
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "INVESTIGATING",
                            "SAR_FILED",
                            "CLOSED"
                        ]
                    },
                    "reason": {
                        "maxLength": 1000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "disposition": {
                        "type": "string",
                        "enum": [
                            "FALSE_POSITIVE",
                            "TRUE_POSITIVE",
                            "NO_FURTHER_ACTION"
                        ]
                    }
                }
            },
...
            "ActionRequest": {
                "required": [
                    "action"
                ],
                "type": "object",
                "properties": {
                    "action": {
                        "type": "string",
                        "enum": [
                            "START_REVIEW",
                            "CLOSE",
                            "ESCALATE",
                            "REOPEN"
                        ]
                    },
                    "disposition": {
                        "type": "string",
                        "enum": [
                            "FALSE_POSITIVE",
                            "TRUE_POSITIVE",
                            "NO_FURTHER_ACTION"
                        ]
                    },
                    "reason": {
                        "maxLength": 1000,
                        "minLength": 0,
                        "type": "string"
                    },
                    "caseId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "AlertDetail": {
                "type": "object",
                "properties": {
                    "summary": {
                        "$ref": "#/components/schemas/AlertSummary"
                    },
                    "explanation": {
                        "type": "string"
                    },
                    "ruleDetails": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/RuleDetail"
                        }
                    },
                    "disposition": {
                        "type": "string"
                    },
                    "dispositionReason": {
                        "type": "string"
                    },
                    "disposedBy": {
                        "type": "string"
                    },
                    "disposedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "caseRef": {
                        "type": "string"
                    },
                    "customer": {
                        "$ref": "#/components/schemas/CustomerView"
                    },
                    "evidence": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EvidenceTxn"
                        }
                    }
                }
            },
...
            "RuleDetail": {
                "type": "object",
                "properties": {
                    "severity": {
                        "type": "number",
                        "format": "double"
                    },
                    "hits": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "ruleVersion": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "weight": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "explanation": {
                        "type": "string"
                    },
                    "lastTxnRef": {
                        "type": "string"
                    }
                }
            },
...
            "RuleConfigHistory": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "ruleCode": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "weight": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "params": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    },
                    "changedBy": {
                        "type": "string"
                    },
                    "changedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "PageResponseAlertSummary": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertSummary"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            }
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}